# api/docs/swagger.yaml
info:
  contact: {}
paths:
...
  /api/v1/rocinante/bootstrap:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: options
        in: query
        name: options
        type: string
      - description: url
        in: query
        name: url
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/cmd:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: args
        in: query
        name: args
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/limits:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: args
        in: query
        name: args
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/list:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: options
        in: query
        name: options
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/pkg:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: args
        in: query
        name: args
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/service:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: ARGS
        in: query
        name: ARGS
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/sysctl:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: args
        in: query
        name: args
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/sysrc:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: args
        in: query
        name: args
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/template:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: options
        in: query
        name: options
        type: string
      - description: action
        in: query
        name: action
        type: string
      - description: template
        in: query
        name: template
        type: string
      - description: args
        in: query
        name: args
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/update:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: args
        in: query
        name: args
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/upgrade:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: args
        in: query
        name: args
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/verify:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: options
        in: query
        name: options
        type: string
      - description: template
        in: query
        name: template
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/zfs:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: args
        in: query
        name: args
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
  /api/v1/rocinante/zpool:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: args
        in: query
        name: args
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - rocinante
swagger: "2.0"